[FEAT] PhotoDetail: 사진 데이터 연결 및 사진 크기에 맞게 뷰 크기 조정 #1

+++ PhotoExplorer/PhotoExplorer/PhotoSearchResponse.swift
@@ -15,8 +15,8 @@ struct PhotoSearchResponse: Decodable {
 
 struct Photo: Decodable {
     let id: String
-    let width: Int
-    let height: Int
+    let width: CGFloat
+    let height: CGFloat
     let color: String
     let likes: Int
     let urls: PhotoURLs
